+++ .changeset/giant-penguins-hang.md
@@ -1,0 +1,5 @@
+---
+"@uploadthing/shared": patch
+---
+
+chore: move some types to `@uploadthing/shared`
