--- packages/uploadthing/src/constants.ts
@@ -1,1 +1,0 @@
-export { version as UPLOADTHING_VERSION } from "../package.json";
