+++ packages/uploadthing/src/express.ts
@@ -7,7 +7,7 @@ import type {
 import type { Json } from "@uploadthing/shared";
 import { getStatusCodeFromError, UploadThingError } from "@uploadthing/shared";
 
-import { UPLOADTHING_VERSION } from "./constants";
+import { UPLOADTHING_VERSION } from "./internal/constants";
 import { formatError } from "./internal/error-formatter";
 import type { RouterWithConfig } from "./internal/handler";
 import {
