+++ packages/uploadthing/src/h3.ts
@@ -10,7 +10,7 @@ import {
 import type { Json } from "@uploadthing/shared";
 import { getStatusCodeFromError, UploadThingError } from "@uploadthing/shared";
 
-import { UPLOADTHING_VERSION } from "./constants";
+import { UPLOADTHING_VERSION } from "./internal/constants";
 import { defaultErrorFormatter } from "./internal/error-formatter";
 import {
   buildPermissionsInfoHandler,
