+++ packages/uploadthing/src/internal/constants.ts
@@ -1,0 +1,1 @@
+export { version as UPLOADTHING_VERSION } from "../../package.json";
