+++ packages/uploadthing/src/internal/handler.ts
@@ -18,7 +18,7 @@ import type {
   UploadedFile,
 } from "@uploadthing/shared";
 
-import { UPLOADTHING_VERSION } from "../constants";
+import { UPLOADTHING_VERSION } from "./constants";
 import { conditionalDevServer } from "./dev-hook";
 import { getFullApiUrl } from "./get-full-api-url";
 import type { LogLevel } from "./logger";
