+++ packages/uploadthing/test/__test-helpers.ts
@@ -2,7 +2,7 @@ import { beforeEach, vi } from "vitest";
 
 import type { FetchEsque } from "@uploadthing/shared";
 
-import { UPLOADTHING_VERSION } from "../src/constants";
+import { UPLOADTHING_VERSION } from "../src/internal/constants";
 import type { PSPResponse } from "../src/internal/handler";
 import type { ActionType } from "../src/server";
